+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v3.1.0
  1. Added full support to OneView Rest API version 300 for the hardware variants C7000 and Synergy to the already existing features:
+   - Logical Interconnect
    - Uplink Set
    - Volume attachment